+++ api/src/test/java/com/example/echo_api/modules/profile/controller/ProfileViewControllerIT.java
@@ -11,12 +11,12 @@ import org.springframework.core.ParameterizedTypeReference;
 import org.springframework.http.HttpStatus;
 
 import com.example.echo_api.config.ErrorMessageConfig;
+import com.example.echo_api.exception.ErrorResponse;
 import com.example.echo_api.modules.profile.dto.response.ProfileDTO;
 import com.example.echo_api.modules.profile.dto.response.SimplifiedProfileDTO;
 import com.example.echo_api.modules.profile.entity.Follow;
 import com.example.echo_api.modules.profile.repository.FollowRepository;
 import com.example.echo_api.shared.constant.ApiRoutes;
-import com.example.echo_api.shared.dto.ErrorDTO;
 import com.example.echo_api.shared.pagination.PageDTO;
 import com.example.echo_api.testing.support.AbstractIntegrationTest;
 
@@ -99,7 +99,7 @@ class ProfileViewControllerIT extends AbstractIntegrationTest {
         // api: GET /api/v1/profile/{username} ==> 404 Not Found : ErrorDTO
         String nonExistingUsername = "i_dont_exist";
 
-        ErrorDTO expected = new ErrorDTO(
+        ErrorResponse expected = new ErrorResponse(
             HttpStatus.NOT_FOUND,
             ErrorMessageConfig.NotFound.RESOURCE_NOT_FOUND,
             null);
@@ -108,7 +108,7 @@ class ProfileViewControllerIT extends AbstractIntegrationTest {
             .uri(BY_USERNAME_PATH, nonExistingUsername)
             .exchange()
             .expectStatus().isNotFound()
-            .expectBody(ErrorDTO.class).isEqualTo(expected);
+            .expectBody(ErrorResponse.class).isEqualTo(expected);
     }
 
     @Test
@@ -137,7 +137,7 @@ class ProfileViewControllerIT extends AbstractIntegrationTest {
         // api: GET /api/v1/profile/{id}/followers ==> 404 Not Found : ErrorDTO
         UUID nonExistingProfileId = UUID.randomUUID();
 
-        ErrorDTO expected = new ErrorDTO(
+        ErrorResponse expected = new ErrorResponse(
             HttpStatus.NOT_FOUND,
             ErrorMessageConfig.NotFound.RESOURCE_NOT_FOUND,
             null);
@@ -146,7 +146,7 @@ class ProfileViewControllerIT extends AbstractIntegrationTest {
             .uri(FOLLOWERS_BY_ID_PATH, nonExistingProfileId)
             .exchange()
             .expectStatus().isNotFound()
-            .expectBody(ErrorDTO.class).isEqualTo(expected);
+            .expectBody(ErrorResponse.class).isEqualTo(expected);
     }
 
     @Test
@@ -175,7 +175,7 @@ class ProfileViewControllerIT extends AbstractIntegrationTest {
         // api: GET /api/v1/profile/{id}/following ==> 404 Not Found : ErrorDTO
         UUID nonExistingProfileId = UUID.randomUUID();
 
-        ErrorDTO expected = new ErrorDTO(
+        ErrorResponse expected = new ErrorResponse(
             HttpStatus.NOT_FOUND,
             ErrorMessageConfig.NotFound.RESOURCE_NOT_FOUND,
             null);
@@ -184,7 +184,7 @@ class ProfileViewControllerIT extends AbstractIntegrationTest {
             .uri(FOLLOWING_BY_ID_PATH, nonExistingProfileId)
             .exchange()
             .expectStatus().isNotFound()
-            .expectBody(ErrorDTO.class).isEqualTo(expected);
+            .expectBody(ErrorResponse.class).isEqualTo(expected);
     }
 
 }
